rip out the android searchbar in favor of react-native-paper's searchbar
this breaks … oh, every other view with a search bar, but it's worth it
